pyproject.toml
[tool.poetry]
name = "anime-backend"
version = "0.1.0"
description = ""
authors = ["J0hns0n-Wang <johnsonwang529@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.5"
graphene-django = "^3.0.0"
django-filter = "^22.1"
django-allauth = "^0.52.0"
django-cors-headers = "^3.13.0"
PyJWT = "^2.6.0"
graphene = "^3.2.1"
django-rest-framework = "^0.1.0"
djangorestframework-simplejwt = "^5.2.2"
python-dotenv = "^0.21.0"
dj-rest-auth = "2.2.5"
snapshottest = "^0.6.0"
dj-database-url = "^2.0.0"
psycopg2-binary = "^2.9.6"
whitenoise = {extras = ["brotli"], version = "^6.4.0"}
gunicorn = "^20.1.0"
chardet = "^5.1.0"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"